requirements.txt
apache-airflow>=2.10.3
pyspark>=3.5.3
pyspark-stubs>=3.0.0.post3
setuptools>=74.1.3
protobuf>=4.25.4
python-dotenv>=1.0.1
docutils>=0.16
Faker>=30.8.1